+++ sub/ext/etkecc/etkecc.go
@@ -2,22 +2,16 @@ package etkecc
 
 import (
 	"github.com/mattevans/postmark-go"
-	"gitlab.com/etke.cc/buscarron/config"
 	"maunium.net/go/mautrix"
+
+	"gitlab.com/etke.cc/buscarron/config"
+	"gitlab.com/etke.cc/buscarron/sub/ext/common"
 )
 
 // Etkecc extension
 type Etkecc struct {
-	v    NetworkValidator
 	pm   EmailSender
 	test bool
-}
-
-// NetworkValidator interface
-type NetworkValidator interface {
-	A(host string) bool
-	CNAME(host string) bool
-	GetBase(domain string) string
 }
 
 // EmailSender interface
@@ -26,16 +20,16 @@ type EmailSender interface {
 }
 
 // New etke.cc extension
-func New(v NetworkValidator, pm EmailSender) *Etkecc {
-	return &Etkecc{v: v, pm: pm}
+func New(pm EmailSender) *Etkecc {
+	return &Etkecc{pm: pm}
 }
 
 // Execute extension
-func (e *Etkecc) Execute(form *config.Form, data map[string]string) (string, []*mautrix.ReqUploadMedia) {
-	return parseOrder(form.Name, data, e.v, e.pm, e.test)
+func (e *Etkecc) Execute(v common.Validator, form *config.Form, data map[string]string) (string, []*mautrix.ReqUploadMedia) {
+	return parseOrder(form.Name, data, v, e.pm, e.test)
 }
 
-func parseOrder(name string, data map[string]string, v NetworkValidator, pm EmailSender, test bool) (string, []*mautrix.ReqUploadMedia) {
+func parseOrder(name string, data map[string]string, v common.Validator, pm EmailSender, test bool) (string, []*mautrix.ReqUploadMedia) {
 	o := &order{
 		name:  name,
 		data:  data,