+++ sub/handler.go
@@ -30,7 +30,7 @@ type EmailSender interface {
 
 // Validator interface
 type Validator interface {
-	Domain(string, bool) bool
+	Domain(string) bool
 	DomainString(string) bool
 	Email(string) bool
 	A(string) bool
@@ -47,7 +47,7 @@ type Handler struct {
 	forms       map[string]*config.Form
 	log         *logger.Logger
 	ext         map[string]ext.Extension
-	v           Validator
+	vs          map[string]Validator
 }
 
 const redirect = "<html><head><title>Redirecting...</title><meta http-equiv=\"Refresh\" content=\"0; url='{{ .URL }}'\" /></head><body>Redirecting to <a href='{{ .URL }}'>{{ .URL }}</a>..."
@@ -60,15 +60,15 @@ var (
 )
 
 // NewHandler creates new HTTP forms handler
-func NewHandler(forms map[string]*config.Form, v Validator, pm EmailSender, sender Sender, loglevel string) *Handler {
+func NewHandler(forms map[string]*config.Form, vs map[string]Validator, pm EmailSender, sender Sender, loglevel string) *Handler {
 	h := &Handler{
 		redirectTpl: template.Must(template.New("redirect").Parse(redirect)),
 		sanitizer:   bluemonday.StrictPolicy(),
 		sender:      sender,
 		forms:       forms,
 		log:         logger.New("sub.", loglevel),
-		ext:         ext.New(v, pm),
-		v:           v,
+		ext:         ext.New(pm),
+		vs:          vs,
 	}
 
 	return h
@@ -91,6 +91,11 @@ func (h *Handler) POST(rID, name string, r *http.Request) (string, error) {
 		h.log.Warn("submission attempt to the %s form (does not exist)", name)
 		return "", ErrNotFound
 	}
+	v, ok := h.vs[name]
+	if !ok {
+		h.log.Warn("submission attempt to the %s form (validator does not exists)", name)
+		return "", ErrNotFound
+	}
 
 	if err := r.ParseForm(); err != nil {
 		h.log.Error("cannot parse a submission to the %s form: %v", name, err)
@@ -102,12 +107,12 @@ func (h *Handler) POST(rID, name string, r *http.Request) (string, error) {
 		data[key] = strings.TrimSpace(h.sanitizer.Sanitize(r.PostFormValue(key)))
 	}
 
-	if !h.v.Email(data["email"]) {
+	if !v.Email(data["email"]) {
 		h.log.Info("submission to the %s form marked as spam, reason: email", form.Name)
 		return h.redirect(form.Redirect, data), ErrSpam
 	}
 
-	if !h.v.Domain(data["domain"], form.HasDomain) {
+	if !v.Domain(data["domain"]) {
 		h.log.Info("submission to the %s form marked as spam, reason: domain", form.Name)
 		return h.redirect(form.Redirect, data), ErrSpam
 	}
@@ -156,7 +161,8 @@ func (h *Handler) redirect(target string, vars map[string]string) string {
 
 // generate text and files
 func (h *Handler) generate(form *config.Form, data map[string]string) (string, []*mautrix.ReqUploadMedia) {
-	text, medias := h.ext["root"].Execute(form, data)
+	v := h.vs[form.Name]
+	text, medias := h.ext["root"].Execute(v, form, data)
 
 	for _, extension := range form.Extensions {
 		if extension == "" {
@@ -167,7 +173,7 @@ func (h *Handler) generate(form *config.Form, data map[string]string) (string, [
 			continue
 		}
 
-		etext, emedias := e.Execute(form, data)
+		etext, emedias := e.Execute(v, form, data)
 		text += etext
 		medias = append(medias, emedias...)
 	}